% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichedivergence.R
\name{beta_overlap}
\alias{beta_overlap}
\title{Determines response overlap between two beta function curves}
\usage{
beta_overlap(spa, spb)
}
\arguments{
\item{spa}{Data frame containing x and y values for Species A response}

\item{spb}{Data frame containing x and y values for Species B response}
}
\value{
Data frame containing x and y values for the overlap curve between species A and B
}
\description{
Determines response overlap between two beta function curves
}
\examples{
spa <- betaPDF(a = 0, b = 1, alpha = 3, gamma = 1)
spb <- betaPDF(a = 0, b = 1, alpha = 1, gamma = 3)
beta_overlap(spa, spb)
}
