% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaPDF.R
\name{betafunction_max}
\alias{betafunction_max}
\title{Stores Mathematical Expression For 4-Parameter Beta Function}
\usage{
betafunction_max()
}
\value{
An R expression
}
\description{
Stores Mathematical Expression For 4-Parameter Beta Function
}
\examples{
betafunction_max()
}
