% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaPDF.R
\name{betaPDF}
\alias{betaPDF}
\title{Creates X and Y values for a Beta Function Curve}
\usage{
betaPDF(a = 0, b = 1, alpha = 1, gamma = 1, k = 1, interval = 0.001)
}
\arguments{
\item{a}{Lower bound of beta curve}

\item{b}{Upper bound of beta curve}

\item{alpha}{\eqn{\alpha} shape parameter (right skew)}

\item{gamma}{\eqn{\gamma} shape parameter (left skew)}

\item{k}{Scale parameter, usually left at 1 by default}

\item{interval}{Interval length for steps between a and b. By default 0.001}
}
\value{
Data frame containing X and Y values for the desired intervals between a and b
}
\description{
Creates X and Y values for a Beta Function Curve
}
\examples{
a = 0; b = 1
alpha = gamma <- 2
betaPDF(a, b, alpha, gamma)
}
