% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_simulations.R
\name{calculate_nichediv_b1}
\alias{calculate_nichediv_b1}
\title{Niche Divergence Calculations for a given Parameter Space based on the beta function curve}
\usage{
calculate_nichediv_b1(psp)
}
\arguments{
\item{psp}{Parameter space created using the set_parameter_space() function}
}
\value{
A data frame containing niche divergence plane values of Dissimilarity and Exclusivity, along with the parameters of each species response
}
\description{
Niche Divergence Calculations for a given Parameter Space based on the beta function curve
}
\examples{
# Set parameter domains
a = b <- round(1:3, 3)
alpha = gamma <- round(1:3)

# Set parameter space (combination of parameters for any beta function)
par_space <- set_parameter_space(a, b, alpha, gamma)

#Run calc
simulate_nichediv_b1(psp)

}
