% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_simulations.R
\name{calculate_nichediv_beta}
\alias{calculate_nichediv_beta}
\title{Niche Divergence Calculations for a given Parameter Space based on the beta function curve}
\usage{
calculate_nichediv_beta(psp, parallel = FALSE, n = 2)
}
\arguments{
\item{psp}{Parameter space created using the set_parameter_space() function}

\item{parallel}{logical. Change to TRUE if parallel processing is desired}

\item{n}{number of cores to use in parallel processing (see future::plan())}
}
\value{
A data frame containing niche divergence plane values of Dissimilarity and Exclusivity, along with the parameters of each species response
}
\description{
This functions calculates niche divergence indices on all theoretical response curves simulated using the set_parameter_space(). It is a wrapper around two other functions that work in single or parallel processing.
}
\examples{
# Set parameter domains
a = b <- round(1:3, 3)
alpha = gamma <- round(1)

# Set parameter space (combination of parameters for any beta function)
par_space <- set_parameter_space(a, b, alpha, gamma)

#Run calc
calculate_nichediv_beta(psp = par_space) # single core calculation
calculate_nichediv_beta(psp = par_space, parallel = TRUE, n = 2) # parallel calculations
#By default function uses 2 cores, but you can run
#n = parallelly::availableCores()-1
}
