% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_simulations.R
\name{calculate_nichediv_bpar}
\alias{calculate_nichediv_bpar}
\title{(Parallel) Niche Divergence Calculations for a given Parameter Space based on the beta function curve}
\usage{
calculate_nichediv_bpar(psp, n = parallelly::availableCores() - 1)
}
\arguments{
\item{psp}{Parameter space created using the set_parameter_space() function}

\item{n}{number of cores to use in parallel processing (see future::plan())}
}
\value{
A data frame containing niche divergence plane values of Dissimilarity and Exclusivity, along with the parameters of each species response
}
\description{
(Parallel) Niche Divergence Calculations for a given Parameter Space based on the beta function curve
}
\examples{
# Set parameter domains
a = b <- round(1:3, 3)
alpha = gamma <- round(1)

# Set parameter space (combination of parameters for any beta function)
par_space <- set_parameter_space(a, b, alpha, gamma)

#Run calc
calculate_nichediv_bpar(psp = par_space)
}
