% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{create_multi_sa}
\alias{create_multi_sa}
\title{Create study areas for multiple species occurrences}
\usage{
create_multi_sa(
  occs_list,
  buffer = 1e+05,
  bg_area = NULL,
  taxa_field = NULL,
  desired_crs = NULL
)
}
\arguments{
\item{occs_list}{list or SpatVector List of species occurrence spatial objects or a single SpatVector with a taxa/species identifier column}

\item{buffer}{numeric Buffer distance in meters (default 100000 = 100 km)}

\item{bg_area}{numeric Buffer distance in meters (default 100000 = 100 km)}

\item{taxa_field}{character Name of the taxa/species identifier field in single SpatVector. Default NULL.}

\item{desired_crs}{character EPSG code or CRS string to reproject all SpatVectors to (optional)}
}
\value{
named list of study area proxies for each species
}
\description{
Create study areas for multiple species occurrences
}
