% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{create_sa}
\alias{create_sa}
\title{Create study area proxy using buffers around occurrences}
\usage{
create_sa(occs, buffer = 1e+05, bg_area = NULL)
}
\arguments{
\item{occs}{list List of species occurrence spatial objects}

\item{buffer}{numeric Buffer distance in meters (default 100000 = 100 km)}

\item{bg_area}{numeric Buffer distance in meters (default 100000 = 100 km)}
}
\value{
list containing study area proxy based on buffer around occurrence points
}
\description{
Create study area proxy using buffers around occurrences
}
