% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_mx.R
\name{default_regularization}
\alias{default_regularization}
\title{Default Regularization for Maxent-like Models}
\usage{
default_regularization(p, m)
}
\arguments{
\item{p}{Presence/background vector}

\item{m}{Model matrix}
}
\description{
Default Regularization for Maxent-like Models
}
