% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{extract_environmental_data}
\alias{extract_environmental_data}
\title{Load environmental rasters, extract raster values for occurrences and background, remove incomplete cases}
\usage{
extract_environmental_data(ras, occ, bgp)
}
\arguments{
\item{ras}{character Vector of paths to environmental raster files or SpatRaster}

\item{occ}{SpatVector or object convertible to SpatVector Occurrence points}

\item{bgp}{SpatVector or object convertible to SpatVector Background points}
}
\value{
list with occurrence_data and background_data data frames
}
\description{
Load environmental rasters, extract raster values for occurrences and background, remove incomplete cases
}
