% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_ndp_functions.R
\name{format_vsp}
\alias{format_vsp}
\title{Format parameter list for virtualspecies::formatFunctions}
\usage{
format_vsp(param_list)
}
\arguments{
\item{param_list}{A named list of parameter rows (each 1-row tibble)}
}
\value{
A list formatted for generateSpFromFun()
}
\description{
Format parameter list for virtualspecies::formatFunctions
}
