% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{get_response_curve}
\alias{get_response_curve}
\title{Generate Response Curve Data from Model}
\usage{
get_response_curve(model, variable, range = NULL, n = 100)
}
\arguments{
\item{model}{A fitted model object (e.g., from glmnet_mx)}

\item{variable}{Character, name of the variable to generate the curve for}

\item{range}{Numeric vector of length 2, the range of the variable}

\item{n}{Numeric, number of points to generate}
}
\description{
Generate Response Curve Data from Model
}
