% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_mx.R
\name{glmnet_mx}
\alias{glmnet_mx}
\title{Maxent-like glmnet models}
\usage{
glmnet_mx(
  p,
  data,
  f,
  regmult = 1,
  regfun = maxnet::maxnet.default.regularization,
  addsamplestobackground = TRUE,
  weights = NULL,
  ...
)
}
\arguments{
\item{p}{A vector of binary presence-background labels (1 presence, 0 background)}

\item{data}{A data.frame containing predictor variables}

\item{f}{A formula for the model}

\item{regmult}{Regularization multiplier}

\item{regfun}{Function to calculate regularization penalties}

\item{addsamplestobackground}{Logical, whether to add presence points to background}

\item{weights}{Numeric vector of weights}

\item{...}{Additional arguments to glmnet}
}
\description{
Maxent-like glmnet models
}
