% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{maxent_ndp}
\alias{maxent_ndp}
\title{Calculate niche divergence plane indices based on maxent response curves}
\usage{
maxent_ndp(mx_sp1, mx_sp2, cut_off = 0.95)
}
\arguments{
\item{mx_sp1}{maxent model object 1 (EMNTools::enmtools.maxent output)}

\item{mx_sp2}{maxent model object 2 (EMNTools::enmtools.maxent output)}

\item{cut_off}{numeric Proportion cutoff to trim response curve tails (default 0.95)}
}
\value{
list containing:
\itemize{
\item indices: data.frame with dissimilarity and exclusivity indices
\item response_curves: combined trimmed response curves of the two species
}
}
\description{
Calculate niche divergence plane indices based on maxent response curves
}
