% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet2.R
\name{maxnet2}
\alias{maxnet2}
\title{Modified Maxent Model Builder Using glmnet}
\usage{
maxnet2(
  p,
  data,
  f = maxnet.formula(p, data),
  regmult = 1,
  regfun = maxnet.default.regularization,
  addsamplestobackground = TRUE,
  ...
)
}
\arguments{
\item{p}{ a vector of 1 (for presence) or 0 (for background). }

\item{data}{ a matrix or data frame of predictor variables. }

\item{f}{ a formula to determine the features to be used. }

\item{regmult}{ a constant to adjust regularization. }

\item{regfun}{ a function to compute regularization constant for each feature. }

\item{addsamplestobackground}{ if T, add to the background any presence sample that is not already there. }

\item{...}{Additional parameters passed to \code{glmnet}.}
}
\value{
An object of class \code{maxnet} containing the fitted model and metadata.
}
\description{
This function is a modified version of the original maxnet function from the
\pkg{maxnet} package, adapted for specific modeling needs.
For details on the original function, see \code{\link[maxnet]{maxnet}}.
}
\details{
This version modifies the fitting process by adding unique presence samples to the background,
setting controlled regularization adjustments, and calculating model entropy and penalty details.
}
