% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp_tests.R
\name{ndp_analysis}
\alias{ndp_analysis}
\title{Niche Divergence Plane Analysis for Two Species}
\usage{
ndp_analysis(data_a, data_b, formula, variables)
}
\arguments{
\item{data_a}{Presence-background data for Species A}

\item{data_b}{Presence-background data for Species B}

\item{formula}{Formula for the glmnet_mx models}

\item{variables}{Vector of variable names to analyze}
}
\description{
Niche Divergence Plane Analysis for Two Species
}
