% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndp_tests.R
\name{ndp_bootstrap}
\alias{ndp_bootstrap}
\title{Bootstrap Niche Divergence Plane}
\usage{
ndp_bootstrap(data_a, data_b, formula, variables, n_iter = 100)
}
\arguments{
\item{data_a}{Presence-background data for Species A}

\item{data_b}{Presence-background data for Species B}

\item{formula}{Formula for the glmnet_mx models}

\item{variables}{Vector of variable names to analyze}

\item{n_iter}{Number of bootstrap iterations}
}
\description{
Bootstrap Niche Divergence Plane
}
