% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichedivergence.R
\name{niche_diss}
\alias{niche_diss}
\title{Calculate Niche Dissimilarity Index for the Niche Divergence Plane}
\usage{
niche_diss(spa, spb, ov)
}
\arguments{
\item{spa}{Data frame containing x and y values for Species A response}

\item{spb}{Data frame containing x and y values for Species B response}

\item{ov}{Data frame containing x and y values for Overlap between A and B}
}
\value{
Numeric Niche Dissimilarity value
}
\description{
Calculate Niche Dissimilarity Index for the Niche Divergence Plane
}
\examples{
spa <- betaPDF(a = 0, b = 1, alpha = 3, gamma = 1)
spb <- betaPDF(a = 0, b = 1, alpha = 1, gamma = 3)
ov <- beta_overlap(spa, spb)
niche_diss(spa, spb, ov)
}
