% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichedivergence.R
\name{niche_excl}
\alias{niche_excl}
\title{Calculate Niche Exclusivity Index for the Niche Divergence Plane}
\usage{
niche_excl(spa, spb)
}
\arguments{
\item{spa}{Data frame containing x and y values for Species A response}

\item{spb}{Data frame containing x and y values for Species B response}
}
\value{
Numeric Niche Exclusivity value
}
\description{
Calculate Niche Exclusivity Index for the Niche Divergence Plane
}
\examples{
spa <- betaPDF(a = 0, b = 1, alpha = 3, gamma = 1)
spb <- betaPDF(a = 0, b = 1, alpha = 1, gamma = 3)
niche_excl(spa, spb)
}
