% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_ndp_functions.R
\name{normalize_suitability}
\alias{normalize_suitability}
\title{Normalize Suitability Raster from Virtual Species}
\usage{
normalize_suitability(vsp)
}
\arguments{
\item{vsp}{A \code{virtualspecies} object generated with \code{generateSpFromFun()}.}
}
\value{
A \code{SpatRaster} object with normalized values.
}
\description{
Normalizes a habitat suitability raster from a \code{virtualspecies} object so that all pixel values sum to 1.
}
