% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{old_thin_records}
\alias{old_thin_records}
\title{Thin occurrence records based on pairwise distances}
\usage{
old_thin_records(x, thin.par, reps = 10)
}
\arguments{
\item{x}{A matrix, data.frame (lon/lat), or \code{SpatVector} of coordinates.}

\item{thin.par}{Minimum pairwise distance in kilometers.}

\item{reps}{Number of thinning replicates to perform (default = 10).}
}
\value{
A matrix or data.frame of thinned coordinates (longitude, latitude).
}
\description{
Applies spatial thinning to remove points closer than a specified minimum distance. Assumes geographic coordinates (EPSG:4326) unless otherwise specified. Projects to EPSG:4326 if necessary.
}
\seealso{
Other thinning: 
\code{\link{old_adaptive_thin}()}
}
\concept{thinning}
