% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ndp}
\alias{plot_ndp}
\title{Plot Niche Divergence Plane}
\usage{
plot_ndp(indices, labels = NULL, main = "Niche Divergence Plane")
}
\arguments{
\item{indices}{Data frame with exclusivity and dissimilarity columns}

\item{labels}{Optional vector of labels for the points}

\item{main}{Title for the plot}
}
\description{
Plot Niche Divergence Plane
}
