% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_response_curves}
\alias{plot_response_curves}
\title{Plot Response Curves for Two Species}
\usage{
plot_response_curves(
  spa,
  spb,
  main = "Response Curves",
  xlab = "Environmental Gradient",
  col_a = "black",
  col_b = "blue"
)
}
\arguments{
\item{spa}{Data frame for Species A (x, y)}

\item{spb}{Data frame for Species B (x, y)}

\item{main}{Title for the plot}

\item{xlab}{Label for x axis}

\item{col_a}{Color for Species A}

\item{col_b}{Color for Species B}
}
\description{
Plot Response Curves for Two Species
}
