% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_mx.R
\name{predict.glmnet_mx}
\alias{predict.glmnet_mx}
\title{Predict method for glmnet_mx objects}
\usage{
\method{predict}{glmnet_mx}(
  object,
  newdata,
  type = c("link", "response", "exponent", "cloglog"),
  clamp = TRUE,
  ...
)
}
\arguments{
\item{object}{A glmnet_mx model object}

\item{newdata}{Data frame of new environmental data}

\item{type}{Type of prediction: "link", "response", "exponent", "cloglog"}

\item{clamp}{Logical, whether to clamp predictors to the range used during training}

\item{...}{Additional arguments}
}
\description{
Predict method for glmnet_mx objects
}
