% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{read_point_shapefiles}
\alias{read_point_shapefiles}
\title{Read and validate list of shapefiles containing points}
\usage{
read_point_shapefiles(file_paths)
}
\arguments{
\item{file_paths}{character vector of file paths to shapefiles}
}
\value{
list of SpatVectors with point geometries
}
\description{
Read and validate list of shapefiles containing points
}
