% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{run_maxent_ndp}
\alias{run_maxent_ndp}
\title{Run maxent analysis for one or more environmental variables}
\usage{
run_maxent_ndp(
  envdata,
  species_occurrences,
  species_backgrounds,
  study_areas,
  variables = "all"
)
}
\arguments{
\item{envdata}{SpatRaster Environmental raster data}

\item{species_occurrences}{list List of species-specific occurrence SpatVectors}

\item{species_backgrounds}{list List of species-specific background SpatVectors}

\item{study_areas}{list List of study areas (SpatVectors) corresponding to species}

\item{variables}{character or numeric "all" (default) or a subset of variable names or indices}
}
\value{
list containing ind_res and plot_res result objects
}
\description{
Run maxent analysis for one or more environmental variables
}
