% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_ndp_functions.R
\name{same_crs}
\alias{same_crs}
\title{Check if all SpatVectors have the same CRS}
\usage{
same_crs(svs)
}
\arguments{
\item{svs}{list of SpatVectors}
}
\value{
logical TRUE if all have the same CRS, FALSE otherwise
}
\description{
Check if all SpatVectors have the same CRS
}
