% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_ndp_functions.R
\name{sample_params}
\alias{sample_params}
\title{Randomly sample one parameter set per variable}
\usage{
sample_params(param_grids, strategy = "random")
}
\arguments{
\item{param_grids}{Named list of parameter data frames (from define_vsp_param_space)}

\item{strategy}{Sampling strategy: "random" or "first"}
}
\value{
Named list of sampled rows as named vectors
}
\description{
Randomly sample one parameter set per variable
}
