% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameter_space.R
\name{set_parameter_space}
\alias{set_parameter_space}
\title{Create Parameter Space for Niche Divergence Plane Simulations}
\usage{
set_parameter_space(a, b, alpha, gamma)
}
\arguments{
\item{a}{Numerical vector containing set of a values for beta function}

\item{b}{Numerical vector containing set of b values for beta function}

\item{alpha}{Numerical vector containing set of alpha values for beta function}

\item{gamma}{Numerical vector containing set of gamma values for beta function}
}
\value{
Data frame containing all combinations of parameters a, b, alpha, and gamma, where a < b
}
\description{
Create Parameter Space for Niche Divergence Plane Simulations
}
\examples{
a = b <- round(seq(0, 1, 0.1), 3)
alpha = gamma <- round(c(0.001, 0.01, seq(0.1, 1, 0.2), seq(2, 10, 2), 100, 1000), 3)
set_parameter_space(a, b, alpha, gamma)
}
