% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_ndp_functions.R
\name{simulate_vsp}
\alias{simulate_vsp}
\title{Simulate two virtual species and compute similarity metrics}
\usage{
simulate_vsp(params_A, params_B, env_stack)
}
\arguments{
\item{params_A}{Formatted functions list for species A}

\item{params_B}{Formatted functions list for species B}

\item{env_stack}{SpatRaster stack used to generate species}
}
\value{
List with rasters and similarity indices
}
\description{
Simulate two virtual species and compute similarity metrics
}
