% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trapz}
\alias{trapz}
\title{Trapezoidal Integration}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{Numerical vector of x values}

\item{y}{Numerical vector of y values}
}
\value{
Numeric area under the curve
}
\description{
Trapezoidal Integration
}
