% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp_ndp_functions.R
\name{vsp_betaFun}
\alias{vsp_betaFun}
\title{Rescaled beta response function (from virtualspecies internal)}
\usage{
vsp_betaFun(x, p1, p2, alpha, gamma)
}
\arguments{
\item{x}{Numeric. Vector of environmental values.}

\item{p1}{Numeric. Lower limit of species tolerance.}

\item{p2}{Numeric. Upper limit of species tolerance.}

\item{alpha}{Numeric. Shape parameter for the rising edge.}

\item{gamma}{Numeric. Shape parameter for the falling edge.}
}
\value{
A numeric vector of habitat suitability values (0 to max k).
}
\description{
This function returns habitat suitability values based on a rescaled beta function.
It is used in virtual species simulations to define response curves along environmental gradients.
}
