% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndplane-package.R
\docType{package}
\name{ndplane-package}
\alias{ndplane}
\alias{ndplane-package}
\title{ndplane: Creates Niche Divergence Plane And Other Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ndplane is a package that allows its user to compare ecological niches following the Niche Divergence Plane methdology and framework. With this we can quantify and classify different niche divergence scenarios across multiple niche axes.
}
\author{
\strong{Maintainer}: Alfredo Ascanio \email{ascanio.alfredoa@gmail.com} (\href{https://orcid.org/0000-0001-9987-7977}{ORCID})

}
\keyword{internal}
