% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maxent_ndp_results.R
\name{plot_response_curves}
\alias{plot_response_curves}
\title{Plot Maxent response curves}
\usage{
plot_response_curves(
  plot_res,
  selected_vars = c("bio1", "bio3", "bio12", "bio14", "Sand", "WC3rdbar"),
  color_palette = c("#000000", "#62A39F")
)
}
\arguments{
\item{plot_res}{Data frame containing response curve data with columns layer, value, source, var}

\item{selected_vars}{Character vector of variables to plot (default c("bio1", "bio3", "bio12", "bio14", "Sand", "WC3rdbar"))}

\item{color_palette}{Named vector of species colors (default c("#000000", "#62A39F"))}
}
\value{
ggplot2 object with faceted line plots of response curves by variable and type
}
\description{
Plot Maxent response curves
}
